refactor code to use click
