+++ vixos/libvirt_utils.py
@@ -8,7 +8,6 @@ def libvirt_connection(uri: str) -> Any:
     conn = libvirt.open(uri)
     if not conn:
         raise RuntimeError(f"Failed to open connection to {uri}")
-
     try:
         yield conn
     finally:
